Rename EdgeNotification to EdgeConsumer. Sync up yaml config with value. Move setEdgeRootRuleChain to edgeService and delete EdgeNotificationService
