+++ application/src/main/java/org/thingsboard/server/service/entitiy/edge/TbEdgeService.java
@@ -25,6 +25,7 @@ import org.thingsboard.server.common.data.id.TenantId;
 import org.thingsboard.server.common.data.rule.RuleChain;
 
 public interface TbEdgeService {
+
     Edge save(Edge edge, RuleChain edgeTemplateRootRuleChain, User user) throws Exception;
 
     void delete(Edge edge, User user);
