+++ application/src/main/java/org/thingsboard/server/service/queue/TbEdgeConsumerService.java
@@ -13,13 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.thingsboard.server.service.edge;
+package org.thingsboard.server.service.queue;
 
-import org.thingsboard.server.common.data.edge.Edge;
-import org.thingsboard.server.common.data.id.RuleChainId;
-import org.thingsboard.server.common.data.id.TenantId;
+import org.springframework.context.ApplicationListener;
+import org.thingsboard.server.queue.discovery.event.PartitionChangeEvent;
 
-public interface EdgeNotificationService {
+public interface TbEdgeConsumerService extends ApplicationListener<PartitionChangeEvent> {
 
-    Edge setEdgeRootRuleChain(TenantId tenantId, Edge edge, RuleChainId ruleChainId) throws Exception;
 }
